lacation refactor
